+++ NEWS.md
@@ -1,6 +1,7 @@
 # argonR 0.2.0.9000
 - add hex
 - add a view argument to `argonPageTemplate()`
+- add analytics argument to `argonPage()`
 - add src_collapsed arg to `argonNavbar()` so that brand logo are displayed on small screens
 - add `argonContact()`
 - add favicon argument to `argonPage()` and update assets folder accordingly